Refactor: Rename view.py to views.py
